backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
flask-bcrypt==1.0.1
PyJWT==2.9.0
mysql-connector-python==9.5.0
python-dotenv==1.2.1

